study_up/requirements.txt
anyio==4.11.0
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
ConfigArgParse==1.7.1
cryptography==46.0.3
firebase_admin==7.1.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.5
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.42.5
locust-cloud==1.29.2
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==25.0
platformdirs==4.5.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
python-engineio==4.12.3
python-socketio==5.15.0
pywin32==311
pyzmq==27.1.0
requests==2.32.4
rsa==4.9.1
simple-websocket==1.1.0
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1
